code refactoring: separate iop-order and iop-profile conversion.
