+++ src/common/histogram.h
@@ -22,6 +22,7 @@
 
 #include "develop/imageop.h"
 #include "develop/pixelpipe.h"
+#include "common/iop_profile.h"
 
 /*
  * histogram region of interest
