+++ src/iop/colorzones.c
@@ -22,6 +22,7 @@
 #endif
 
 #include "bauhaus/bauhaus.h"
+#include "common/iop_profile.h"
 #include "common/colorspaces_inline_conversions.h"
 #include "common/math.h"
 #include "develop/imageop.h"
