+++ src/iop/lut3d.c
@@ -25,6 +25,7 @@
 #include "common/colorspaces.h"
 #include "common/colorspaces_inline_conversions.h"
 #include "common/file_location.h"
+#include "common/iop_profile.h"
 #include "develop/imageop.h"
 #include "dtgtk/button.h"
 #include "gui/gtk.h"
